MeasureDisplayUnit should work with floats

+++ app/src/main/java/de/westnordost/streetcomplete/measure/MeasureDisplayUnit.kt
@@ -8,7 +8,7 @@ import kotlin.math.round
 /** In which unit the measurement is displayed */
 @Serializable
 sealed class MeasureDisplayUnit {
-    abstract fun format(distanceMeters: Double): String
+    abstract fun format(distanceMeters: Float): String
 }
 /** Measurement displayed in meters rounded to the given number of [decimals] */
 @Serializable
@@ -17,12 +17,12 @@ data class MeasureDisplayUnitMeter(val decimals: Int) : MeasureDisplayUnit() {
         require(decimals >= 0)
     }
 
-    override fun format(distanceMeters: Double): String {
+    override fun format(distanceMeters: Float): String {
         return "%.${decimals}f m".format(getRounded(distanceMeters))
     }
 
     /** Returns the given distance in meters rounded to the given precision */
-    fun getRounded(distanceMeters: Double): Double {
+    fun getRounded(distanceMeters: Float): Double {
         val num = 10.0.pow(decimals)
         return round(distanceMeters * num) / num
     }
@@ -34,13 +34,13 @@ data class MeasureDisplayUnitFeetInch(val inchStep: Int) : MeasureDisplayUnit() 
         require(inchStep in arrayOf(1,2,3,4,6,12))
     }
 
-    override fun format(distanceMeters: Double): String {
+    override fun format(distanceMeters: Float): String {
         val (feet, inches) = getRounded(distanceMeters)
         return if (inches < 10) "$feet′ $inches″" else "$feet′$inches″"
     }
 
     /** Returns the given distance in meters as feet + inch */
-    fun getRounded(distanceMeters: Double): Pair<Int, Int> {
+    fun getRounded(distanceMeters: Float): Pair<Int, Int> {
         val distanceFeet = distanceMeters / 0.3048
         var feet = floor(distanceFeet).toInt()
         val inches = (distanceFeet - feet) * 12
